+++ example/lib/main.dart
@@ -272,8 +272,7 @@ class HomePageState extends State<HomePage> {
                           onTap: () async {
                             showConfirmDialogCustom(
                               context,
-                              dialogAnimation:
-                                  DialogAnimation.SLIDE_RIGHT_LEFT,
+                              dialogAnimation: DialogAnimation.SLIDE_RIGHT_LEFT,
                               title: "Do you want to logout from the app?",
                               dialogType: DialogType.CONFIRMATION,
                               centerImage:
@@ -294,8 +293,7 @@ class HomePageState extends State<HomePage> {
                           onTap: () {
                             showConfirmDialogCustom(
                               context,
-                              dialogAnimation:
-                                  DialogAnimation.SLIDE_BOTTOM_TOP,
+                              dialogAnimation: DialogAnimation.SLIDE_BOTTOM_TOP,
                               title: "Do you want to update this item?",
                               dialogType: DialogType.UPDATE,
                               onAccept: (_) {
@@ -309,8 +307,7 @@ class HomePageState extends State<HomePage> {
                           onTap: () {
                             showConfirmDialogCustom(
                               context,
-                              dialogAnimation:
-                                  DialogAnimation.SLIDE_LEFT_RIGHT,
+                              dialogAnimation: DialogAnimation.SLIDE_LEFT_RIGHT,
                               title: "Delete 89 files permanent?",
                               dialogType: DialogType.DELETE,
                               onAccept: (_) {
@@ -326,8 +323,7 @@ class HomePageState extends State<HomePage> {
                               context,
                               title: "Do you want to add this item?",
                               dialogType: DialogType.ADD,
-                              dialogAnimation:
-                                  DialogAnimation.SLIDE_TOP_BOTTOM,
+                              dialogAnimation: DialogAnimation.SLIDE_TOP_BOTTOM,
                               onAccept: (_) {
                                 snackBar(context, title: 'Added');
                               },
@@ -381,8 +377,7 @@ class HomePageState extends State<HomePage> {
                             showInDialog(
                               context,
                               builder: (_) => dialogWidget(),
-                              dialogAnimation:
-                                  DialogAnimation.SLIDE_TOP_BOTTOM,
+                              dialogAnimation: DialogAnimation.SLIDE_TOP_BOTTOM,
                             );
                           },
                         ),
@@ -392,8 +387,7 @@ class HomePageState extends State<HomePage> {
                             showInDialog(
                               context,
                               builder: (_) => dialogWidget(),
-                              dialogAnimation:
-                                  DialogAnimation.SLIDE_BOTTOM_TOP,
+                              dialogAnimation: DialogAnimation.SLIDE_BOTTOM_TOP,
                             );
                           },
                         ),
@@ -403,8 +397,7 @@ class HomePageState extends State<HomePage> {
                             showInDialog(
                               context,
                               builder: (_) => dialogWidget(),
-                              dialogAnimation:
-                                  DialogAnimation.SLIDE_LEFT_RIGHT,
+                              dialogAnimation: DialogAnimation.SLIDE_LEFT_RIGHT,
                             );
                           },
                         ),
@@ -414,8 +407,7 @@ class HomePageState extends State<HomePage> {
                             showInDialog(
                               context,
                               builder: (_) => dialogWidget(),
-                              dialogAnimation:
-                                  DialogAnimation.SLIDE_RIGHT_LEFT,
+                              dialogAnimation: DialogAnimation.SLIDE_RIGHT_LEFT,
                             );
                           },
                         ),